создан LogMixin, который переопределяет init
